feat(39): implement new Calendar component with event management and improved UI
